feat: Integrate coffi-modern and elfio-modern as submodules and update dependencies
Added new submodules for coffi-modern and elfio-modern to the project, enhancing support for PE and ELF file editing. Updated the CMake configuration to include these libraries, ensuring they are properly linked during the build process. Introduced a new header file, coffi_adapter.hpp, to bridge the result types of coffi-modern and elfio-modern with tl::expected, facilitating error handling. Refactored existing PE and ELF editor implementations to utilize the new libraries, improving maintainability and performance. Updated related tests to ensure compatibility with the new dependencies.

+++ loader/src/X86_32StubEmitter.cpp
@@ -57,7 +57,7 @@ public:
         c.push_back(0x83); c.push_back(0xEC); c.push_back(0x20);  // sub esp, 32
         for (int i = 0; i < 8; ++i) {
             c.push_back(0x89);
-            c.push_back(0x44 | ((i & 7) << 3));  // ModRM [esp+disp8], reg
+            c.push_back(static_cast<uint8_t>(0x44 | ((i & 7) << 3)));  // ModRM [esp+disp8], reg
             c.push_back(0x24);                    // SIB base=esp
             c.push_back(static_cast<uint8_t>(i * 4));
         }
